pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ibridaDB"
version = "0.2.0"
description = "Modular, reproducible database system for biodiversity observations from iNaturalist and expert datasets"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Caleb", email = "caleb@example.com"}
]
maintainers = [
    {name = "Caleb", email = "caleb@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11", 
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.10"
dependencies = [
    # Database connectivity
    "psycopg2-binary>=2.9.9",
    "SQLAlchemy>=2.0.30",
    "PyMySQL>=1.1.1",
    
    # Data processing
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "tqdm>=4.66.4",
    "polars>=0.20.0",
    
    # Image processing for anthophila manifest
    "Pillow>=10.0.0",
    "ImageHash>=4.3.1",
    
    # Async and Redis (if needed)
    "redis>=5.0.4",
    "async-timeout>=4.0.3",
    "greenlet>=3.0.3",
    
    # Cryptography
    "cryptography>=42.0.8",
    "cffi>=1.16.0",
    "pycparser>=2.22",
    
    # Date/time utilities
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",
    "tzdata>=2024.1",
    
    # Type annotations
    "typing_extensions>=4.11.0",
    
    # Utilities
    "six>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

anthophila = [
    "Pillow>=10.0.0",
    "ImageHash>=4.3.1",
    "rapidfuzz>=3.0.0",  # For fuzzy matching in deduplication
]

export = [
    "h5py>=3.8.0",  # For HDF5 exports if needed
    "scipy>=1.10.0",  # For scientific computations in exports
]

[project.urls]
Homepage = "https://github.com/yourusername/ibridaDB"
Documentation = "https://github.com/yourusername/ibridaDB/blob/main/README.md"
Repository = "https://github.com/yourusername/ibridaDB"
"Bug Tracker" = "https://github.com/yourusername/ibridaDB/issues"

[project.scripts]
build-anthophila-manifest = "scripts.build_anthophila_manifest:main"

[tool.setuptools]
packages = ["models", "scripts", "dbTools"]

[tool.setuptools.package-data]
"*" = ["*.sql", "*.sh", "*.md"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "psycopg2.*",
    "PIL.*",
    "imagehash.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=models",
    "--cov=scripts", 
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "database: marks tests that require database connection",
]